pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-wiki-compiler"
version = "1.0.0"
description = "A Knowledge Graph compiler for LLM-driven development"
authors = [{name = "Your Name", email = "your@email.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    "networkx>=3.0"
]

[project.scripts]
# This exposes the 'wiki-compiler' command in the terminal for installation.
wiki-compiler = "wiki_compiler.main:main"